+++ lib/hadeploy/core/main.py
@@ -86,7 +86,7 @@ def main():
     # Now, build the schema for source validation, by merge of all schema plugin
     theSchema = context.getSchema()
     dump.dumpSchema(context, theSchema)
-    dump.dumpModel(context)
+    #dump.dumpModel(context)
     # And validate against this schema
     schema.validate(context.model[SRC], theSchema)
 
@@ -97,7 +97,7 @@ def main():
     context.buildRemoveTemplate()
     context.builRolesPath()
 
-    #dump.dumpModel(context)
+    dump.dumpModel(context)
 
     context.generatePrivateTemplate()